feat: add ts, eslint, prettier
